Refactor AutoARIMA
